requirements.txt
os
datetime
json
time
pillow
google-cloud-vision
paramiko